moved rename, ensure_directory and jitter to common
Move a few useful functions in downloader used in the backfiller to common
